Firebase protection for event fields messages users

+++ backend/src/routers/messages.py
@@ -1,8 +1,8 @@
 from uuid import UUID
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException, Depends
 from typing import List
 
-from utils.error_handelers import check_no_data
+from utils.firebase import SecurityCheck
 from database.db_engine import engine
 from schema.message import Message, MessageCreate
 from manager import MessageManager
@@ -23,29 +23,40 @@ def get_all_messages():
 
 # Get message by id
 @router.get("/{message_id}", response_model=Message)
-def get_message(message_id: str):
+def get_message(message_id: str, uid: str, authentified_user: str = Depends(SecurityCheck)):
     with engine.begin() as conn:
-        data = MessageManager.get_message_by_id(conn, message_id)
-        check_no_data(data, "message")
-        return data
+        message = MessageManager.get_message_by_id(conn, message_id)
+        if message is None:
+            raise HTTPException(404, "Message not found")
+        else:
+            return message
 
 
 # Create message
 @router.post("/create")
-def create_message(message: MessageCreate):
+def create_message(message: MessageCreate, uid: str, authentified_user: str = Depends(SecurityCheck)):
     with engine.begin() as conn:
-        MessageManager.create_message(conn, message)
+        if authentified_user.id == uid or authentified_user.is_admin:
+            return MessageManager.create_message(conn, message)
+        else:
+            raise HTTPException(409, "Message already exists")
 
 
 # Update message by id
 @router.put("/update/{id}")
-def update_message(message: MessageCreate, id: UUID):
+def update_message(message: MessageCreate, message_id: UUID, uid: str, authentified_user: str = Depends(SecurityCheck)):
     with engine.begin() as conn:
-        MessageManager.update_message(conn, message, id)
+        if authentified_user.id == uid or authentified_user.is_admin:
+            return MessageManager.update_message(conn, message, message_id)
+        else:
+            raise HTTPException(404, "Message not found")
 
 
 # Delete one message by id
 @router.delete("/delete/{message_id}", response_model=bool)
-def delete_event(message_id: str):
+def delete_event(message_id: str, uid: str, authentified_user: str = Depends(SecurityCheck)):
     with engine.begin() as conn:
-        MessageManager.delete_message_by_id(conn, message_id)
+        if authentified_user.id == uid or authentified_user.is_admin:
+            return MessageManager.delete_message_by_id(conn, message_id)
+        else:
+            raise HTTPException(403, "Action not permitted")
